"Update packages in package-lock.json
This commit updates various packages (including @babel/code-frame, @babel/core, and @babel/traverse among others) in package-lock.json. The updates are necessary to ensure compatibility with the latest versions of other packages and keep the project up-to-date with the latest features, improvements, and security patches. The '@esbuild' package has also been upgraded to the latest version for efficient bundling and improved build performance."